Revert "Servings recipe filter added"
