+++ backend/models/recipe.js
@@ -9,9 +9,7 @@ var recipeSchema = new mongoose.Schema({
     servings : String,
     instructions : [
       {type:  String}
-    ],
-    minServings:Number,
-    maxServings:Number
+    ]
 });
 
 module.exports = mongoose.model('recipe', recipeSchema);
