+++ src/Components/RecipeFilter.js
@@ -1,95 +1,27 @@
-import React, { useEffect } from 'react';
-import { withStyles, makeStyles } from '@material-ui/core/styles';
+import React from 'react';
+import { makeStyles } from '@material-ui/core/styles';
 import FormLabel from '@material-ui/core/FormLabel';
 import FormControl from '@material-ui/core/FormControl';
 import FormGroup from '@material-ui/core/FormGroup';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Checkbox from '@material-ui/core/Checkbox';
-import { Slider } from '@material-ui/core';
-import {Col} from 'react-bootstrap'
 
 const useStyles = makeStyles((theme) => ({
   root: {
-    width:200,
-    marginTop:50,
+    display: 'flex',
   },
   formControl: {
-    margin: theme.spacing(3),
+    margin: theme.spacing(1),
   },
 }));
 
-const PrettoSlider = withStyles({
-  root: {
-    color: '#c62828',
-    height: 8,
-    marginTop:50,
-  },
-  thumb: {
-    height: 24,
-    width: 24,
-    backgroundColor: '#fff',
-    border: '2px solid currentColor',
-    marginTop: -8,
-    marginLeft: -12,
-    '&:focus, &:hover, &$active': {
-      boxShadow: 'inherit',
-    },
-  },
-  active: {},
-  valueLabel: {
-    left: 'calc(-50% + 4px)',
-  },
-  track: {
-    height: 8,
-    borderRadius: 4,
-  },
-  rail: {
-    height: 8,
-    borderRadius: 4,
-  },
-})(Slider);
-
 export default function RecipeFilter(props) {
-  const classes = useStyles();
+  //const classes = useStyles();
+ 
 
   const handleChange = (event) => {
     props.setRanges({ ...props.ranges, [event.target.name]: event.target.checked });
   };
-
-  const handleSlider = (event,newVal) =>{
-    props.setSlider(newVal)
-  }
-
-  const allUnchecked = () =>{
-    const { range1,range2,range3,range4,range5,range6 } = props.ranges;
-    if(!range1 && !range2 && !range3 && !range4 && !range5 && !range6)
-      return true;
-    return false;
-  }
-
-
-  useEffect(() =>{
-    let arr = [];
-    const { range1,range2,range3,range4,range5,range6 } = props.ranges;
-    if(range1)
-      arr.push(0,30);
-    if(range2)
-      arr.push(31,60);
-    if(range3)
-      arr.push(61,90);
-    if(range4)
-      arr.push(91,120);
-      if(range5)
-      arr.push(121,150);
-    if(range6)
-      arr.push(150);
-    if(allUnchecked())
-      arr.push(0,30,31,61,90,91,120,121,150)
-    
-    console.log(arr)
-      props.setRangeArr(arr);
-
-  },[props.ranges])
 
   const { range1,range2,range3,range4,range5,range6 } = props.ranges;
 
@@ -124,15 +56,6 @@ export default function RecipeFilter(props) {
           />
         </FormGroup>
       </FormControl>
-      <div className={classes.root}>
-      <FormLabel component="legend">Servings</FormLabel>
-      <Col xs={8} lg={12}>
-      <PrettoSlider valueLabelDisplay="on" aria-label="pretto slider"
-      onChange={handleSlider}
-      min={1}
-      max={50}
-      defaultValue={props.slider} /></Col>
-    </div>
     </div>
   );
 }
